backend/requirements.txt
# FastAPI 核心依赖
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
pydantic-settings==2.6.0

# 数据验证和测试
hypothesis==6.98.0
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-benchmark==4.0.0

# 图片处理
Pillow==10.2.0
qrcode==7.4.2

# 阿里云服务
oss2==2.18.4

# OpenAI SDK (用于 Qwen-VL API)
openai==1.12.0

# HTTP 客户端
httpx==0.26.0
aiohttp==3.9.1

# 工具库
python-dotenv==1.0.0
python-multipart==0.0.6

# 日志和监控
loguru==0.7.2

# MediaPipe (v2.0 功能)
mediapipe==0.10.9
opencv-python==4.9.0.80

# 数据库 (可选)
motor==3.3.2
redis==5.0.1
